requirements.txt
aiofiles==0.7.0
asyncpg==0.23.0
databases==0.4.3
fastapi==0.65.2
Jinja2==3.0.1
psycopg2-binary==2.9.1
python-multipart==0.0.5
PyYAML==6.0
SQLAlchemy==1.3.24
timeout-decorator==0.5.0
uvicorn==0.14.0
docker==5.0.0
six==1.16.0
